refactor: extract parquet projection read plan into its own module
Move-only: `ParquetReadPlan`, `StructFieldAccess`,
`build_projection_read_plan` and the leaf-index/schema-pruning helpers
move from row_filter.rs (2100+ lines) into projection_read_plan.rs.
`PushdownChecker`/`PushdownColumns` become pub(crate) so the new module
can keep using them. No behavior change.
